move schemas to separate folder
