+++ generator/src/redux/schemas/conditions.js
