+++ generator/src/redux/schemas/index.js
@@ -1,0 +1,2 @@
+export * from './conditions'
+export * from './mechanism'
